requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic-settings
